Fixing revisions
