--- src/dispatch/database/revisions/tenant/versions/2022-02-11_1360dfdd8d02.py
@@ -1,42 +1,0 @@
-"""empty message
-
-Revision ID: 1360dfdd8d02
-Revises: 8573e5e95519
-Create Date: 2022-02-11 11:21:26.332073
-
-"""
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision = "1360dfdd8d02"
-down_revision = "8573e5e95519"
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column("alert", sa.Column("orginator", sa.String(), nullable=True))
-    op.add_column("alert", sa.Column("external_link", sa.String(), nullable=True))
-    op.drop_column("plugin_instance", "configuration")
-    op.add_column("source", sa.Column("links", sa.JSON(), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column("source", "links")
-    op.add_column(
-        "plugin_instance",
-        sa.Column(
-            "configuration",
-            postgresql.JSON(astext_type=sa.Text()),
-            autoincrement=False,
-            nullable=True,
-        ),
-    )
-    op.drop_column("alert", "external_link")
-    op.drop_column("alert", "orginator")
-    # ### end Alembic commands ###
