--- src/dispatch/database/revisions/tenant/versions/2022-02-15_1c529dc2aaf3.py
@@ -1,180 +1,0 @@
-"""empty message
-
-Revision ID: 1c529dc2aaf3
-Revises: 1360dfdd8d02
-Create Date: 2022-02-15 14:16:21.394584
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-import sqlalchemy_utils
-
-
-# revision identifiers, used by Alembic.
-revision = "1c529dc2aaf3"
-down_revision = "1360dfdd8d02"
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "source_data_format",
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("name", sa.String(), nullable=True),
-        sa.Column("description", sa.String(), nullable=True),
-        sa.Column("search_vector", sqlalchemy_utils.types.ts_vector.TSVectorType(), nullable=True),
-        sa.Column("project_id", sa.Integer(), nullable=True),
-        sa.ForeignKeyConstraint(["project_id"], ["project.id"], ondelete="CASCADE"),
-        sa.PrimaryKeyConstraint("id"),
-        sa.UniqueConstraint("name", "project_id"),
-    )
-    op.create_index(
-        "source_data_format_search_vector_idx",
-        "source_data_format",
-        ["search_vector"],
-        unique=False,
-        postgresql_using="gin",
-    )
-    op.create_table(
-        "source_environment",
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("name", sa.String(), nullable=True),
-        sa.Column("description", sa.String(), nullable=True),
-        sa.Column("search_vector", sqlalchemy_utils.types.ts_vector.TSVectorType(), nullable=True),
-        sa.Column("project_id", sa.Integer(), nullable=True),
-        sa.ForeignKeyConstraint(["project_id"], ["project.id"], ondelete="CASCADE"),
-        sa.PrimaryKeyConstraint("id"),
-        sa.UniqueConstraint("name", "project_id"),
-    )
-    op.create_index(
-        "source_environment_search_vector_idx",
-        "source_environment",
-        ["search_vector"],
-        unique=False,
-        postgresql_using="gin",
-    )
-    op.create_table(
-        "source_status",
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("name", sa.String(), nullable=True),
-        sa.Column("description", sa.String(), nullable=True),
-        sa.Column("search_vector", sqlalchemy_utils.types.ts_vector.TSVectorType(), nullable=True),
-        sa.Column("project_id", sa.Integer(), nullable=True),
-        sa.ForeignKeyConstraint(["project_id"], ["project.id"], ondelete="CASCADE"),
-        sa.PrimaryKeyConstraint("id"),
-        sa.UniqueConstraint("name", "project_id"),
-    )
-    op.create_index(
-        "source_status_search_vector_idx",
-        "source_status",
-        ["search_vector"],
-        unique=False,
-        postgresql_using="gin",
-    )
-    op.create_table(
-        "source_transport",
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("name", sa.String(), nullable=True),
-        sa.Column("description", sa.String(), nullable=True),
-        sa.Column("search_vector", sqlalchemy_utils.types.ts_vector.TSVectorType(), nullable=True),
-        sa.Column("project_id", sa.Integer(), nullable=True),
-        sa.ForeignKeyConstraint(["project_id"], ["project.id"], ondelete="CASCADE"),
-        sa.PrimaryKeyConstraint("id"),
-        sa.UniqueConstraint("name", "project_id"),
-    )
-    op.create_index(
-        "source_transport_search_vector_idx",
-        "source_transport",
-        ["search_vector"],
-        unique=False,
-        postgresql_using="gin",
-    )
-    op.create_table(
-        "source_type",
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("name", sa.String(), nullable=True),
-        sa.Column("description", sa.String(), nullable=True),
-        sa.Column("search_vector", sqlalchemy_utils.types.ts_vector.TSVectorType(), nullable=True),
-        sa.Column("project_id", sa.Integer(), nullable=True),
-        sa.ForeignKeyConstraint(["project_id"], ["project.id"], ondelete="CASCADE"),
-        sa.PrimaryKeyConstraint("id"),
-        sa.UniqueConstraint("name", "project_id"),
-    )
-    op.create_index(
-        "source_type_search_vector_idx",
-        "source_type",
-        ["search_vector"],
-        unique=False,
-        postgresql_using="gin",
-    )
-    op.add_column("source", sa.Column("source_type_id", sa.Integer(), nullable=True))
-    op.add_column("source", sa.Column("source_status_id", sa.Integer(), nullable=True))
-    op.add_column("source", sa.Column("source_environment_id", sa.Integer(), nullable=True))
-    op.add_column("source", sa.Column("source_data_format_id", sa.Integer(), nullable=True))
-    op.add_column("source", sa.Column("source_transport_id", sa.Integer(), nullable=True))
-    op.add_column("source", sa.Column("project_id", sa.Integer(), nullable=True))
-    op.create_unique_constraint(None, "source", ["name", "project_id"])
-    op.create_foreign_key(None, "source", "source_status", ["source_status_id"], ["id"])
-    op.create_foreign_key(None, "source", "source_type", ["source_type_id"], ["id"])
-    op.create_foreign_key(None, "source", "source_data_format", ["source_data_format_id"], ["id"])
-    op.create_foreign_key(None, "source", "source_transport", ["source_transport_id"], ["id"])
-    op.create_foreign_key(None, "source", "source_environment", ["source_environment_id"], ["id"])
-    op.create_foreign_key(None, "source", "project", ["project_id"], ["id"], ondelete="CASCADE")
-    op.drop_column("source", "transport")
-    op.drop_column("source", "source_type")
-    op.drop_column("source", "data_format")
-    op.drop_column("source", "status")
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column("source", sa.Column("status", sa.VARCHAR(), autoincrement=False, nullable=True))
-    op.add_column(
-        "source", sa.Column("data_format", sa.VARCHAR(), autoincrement=False, nullable=True)
-    )
-    op.add_column(
-        "source", sa.Column("source_type", sa.VARCHAR(), autoincrement=False, nullable=True)
-    )
-    op.add_column(
-        "source", sa.Column("transport", sa.VARCHAR(), autoincrement=False, nullable=True)
-    )
-    op.drop_constraint(None, "source", type_="foreignkey")
-    op.drop_constraint(None, "source", type_="foreignkey")
-    op.drop_constraint(None, "source", type_="foreignkey")
-    op.drop_constraint(None, "source", type_="foreignkey")
-    op.drop_constraint(None, "source", type_="foreignkey")
-    op.drop_constraint(None, "source", type_="foreignkey")
-    op.drop_constraint(None, "source", type_="unique")
-    op.drop_column("source", "project_id")
-    op.drop_column("source", "source_transport_id")
-    op.drop_column("source", "source_data_format_id")
-    op.drop_column("source", "source_environment_id")
-    op.drop_column("source", "source_status_id")
-    op.drop_column("source", "source_type_id")
-    op.drop_index("source_type_search_vector_idx", table_name="source_type", postgresql_using="gin")
-    op.drop_table("source_type")
-    op.drop_index(
-        "source_transport_search_vector_idx", table_name="source_transport", postgresql_using="gin"
-    )
-    op.drop_table("source_transport")
-    op.drop_index(
-        "source_status_search_vector_idx", table_name="source_status", postgresql_using="gin"
-    )
-    op.drop_table("source_status")
-    op.drop_index(
-        "source_environment_search_vector_idx",
-        table_name="source_environment",
-        postgresql_using="gin",
-    )
-    op.drop_table("source_environment")
-    op.drop_index(
-        "source_data_format_search_vector_idx",
-        table_name="source_data_format",
-        postgresql_using="gin",
-    )
-    op.drop_table("source_data_format")
-    # ### end Alembic commands ###
