--- src/dispatch/database/revisions/tenant/versions/2022-02-18_0a1cdd6a48a2.py
@@ -1,30 +1,0 @@
-"""empty message
-
-Revision ID: 0a1cdd6a48a2
-Revises: 64f497e855f4
-Create Date: 2022-02-18 15:28:59.625991
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '0a1cdd6a48a2'
-down_revision = '64f497e855f4'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('query', sa.Column('source_id', sa.Integer(), nullable=True))
-    op.create_foreign_key(None, 'query', 'source', ['source_id'], ['id'])
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'query', type_='foreignkey')
-    op.drop_column('query', 'source_id')
-    # ### end Alembic commands ###
