--- src/dispatch/database/revisions/tenant/versions/2022-02-18_1731ce790ef7.py
@@ -1,42 +1,0 @@
-"""empty message
-
-Revision ID: 1731ce790ef7
-Revises: 1c529dc2aaf3
-Create Date: 2022-02-18 14:21:57.580960
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '1731ce790ef7'
-down_revision = '1c529dc2aaf3'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('assoc_query_tags',
-    sa.Column('query_id', sa.Integer(), nullable=False),
-    sa.Column('tag_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['query_id'], ['query.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['tag_id'], ['tag.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('query_id', 'tag_id')
-    )
-    op.create_table('assoc_query_incidents',
-    sa.Column('query_id', sa.Integer(), nullable=False),
-    sa.Column('incident_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['incident_id'], ['incident.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['query_id'], ['query.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('query_id', 'incident_id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('assoc_query_incidents')
-    op.drop_table('assoc_query_tags')
-    # ### end Alembic commands ###
