--- src/dispatch/database/revisions/tenant/versions/2022-02-18_4ef56fd331bd.py
@@ -1,28 +1,0 @@
-"""empty message
-
-Revision ID: 4ef56fd331bd
-Revises: 0a1cdd6a48a2
-Create Date: 2022-02-18 15:50:24.748130
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '4ef56fd331bd'
-down_revision = '0a1cdd6a48a2'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('query', sa.Column('language', sa.String(), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('query', 'language')
-    # ### end Alembic commands ###
