--- src/dispatch/database/revisions/tenant/versions/2022-02-18_64f497e855f4.py
@@ -1,30 +1,0 @@
-"""empty message
-
-Revision ID: 64f497e855f4
-Revises: 1731ce790ef7
-Create Date: 2022-02-18 15:03:22.471401
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '64f497e855f4'
-down_revision = '1731ce790ef7'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('query', sa.Column('project_id', sa.Integer(), nullable=True))
-    op.create_foreign_key(None, 'query', 'project', ['project_id'], ['id'], ondelete='CASCADE')
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'query', type_='foreignkey')
-    op.drop_column('query', 'project_id')
-    # ### end Alembic commands ###
