+++ src/dispatch/static/dispatch/src/data/source/DetailsTab.vue
@@ -115,8 +115,6 @@ export default {
 
   data() {
     return {
-      labels: ["12am", "3am", "6am", "9am", "12pm", "3pm", "6pm", "9pm"],
-      value: [200, 675, 410, 390, 310, 460, 250, 240],
       headers: [
         {
           text: "Name",
