+++ src/dispatch/static/dispatch/src/data/source/EditBasicInfoTab.vue
@@ -137,6 +137,7 @@ export default {
 
   computed: {
     ...mapFields("source", [
+      "selected.id",
       "selected.name",
       "selected.description",
       "selected.size",
